Define and implement ChessMove class, add chess move test
Define ChessMove class in chessmove.h and implement it in chessmove.cc. Added the
following methods:
- A constructor that takes a ChessPiece* and Position argument
- A constructor that takes a string and ChessBoard argument, the string is used
to construct two Positions, the first postions is used to retrieve the ChessPiece*
from the ChessBoard
- checkValidMoveString() checks whether a string represents two valid positions
on the chessboard (a valid string consists of a file (a-h) and a rank (1-8)
separated by a space, e.g. "e4 f5")
- also some helper methods like charToInt(), isValidRank(), and isValidFile()

Add a chess move test file, chessmove_test.cc. Contains one test that checks the
checkValidMoveString() method