Add getMoves() method to Chess class and various helper methods
Chess.getMoves() takes a ChessPiece argument and calls the a helper method
depending on the piece's type.

Helper methods: getPawnMoves(), getKnightMoves(), getBishopMoves(),
getRookMoves(), getQueenMoves(), getKingMoves().

Pawn, Knight, and King methods all use the pushValidMove() helper method.
This method checks if the position is valid, and pushes the move to a
vector if the move is valid (i.e. move to empty board square or capture).

Rook and Queen methods use pushMovesStraights() helper method. This method
pushes valid moves associated with walking up and down columns, and walking
left and right on rows.

Bishop and Queen methods use pushMovesDiagonals() helper method. This method
pushes valid moves associated with walking up and down diagonals.

Both pushMovesStraights() and pushMovesDiagonals() use a helper method call
pushMove(). This method is different from pushValidMove() in that it doesn't
check if the position argument is valid. When walkin up and down rows,
columns, and diagonals we already know that the positons are valid, so there's
no need to check. Also during these walks we need to stop when we encounter
piece in the way, since any moves beyond that piece are invalid. For this reason
the pushValidMove() returns true if the position is empty and false if the
position is a piece. This return value is used to determine when to stop the walk
since we hit a piece and to stop pushing moves.