Improve the ChessPieceSet class and related tests
Change default constructor to set the positions of the chess pieces to invalid
positions instead of the inital chess boars starting positions.

Added setPiecePositions() method that takse two vectors of positions and uses them
to set the positions of the chess pieces (one vector for the black set, one
for the whte set).

Added another constructor that takes two vectors of positions and uses
setPiecePositions() to set the positions.

Added an internal private setPiecePositions() that just sets the postions of
single player's set.

Added a getPlayerSet() private method that takes a player argument and returns
a pointer to the their chessPiece array.

Changes to the getPieces() methods, both of them now take an optional filterCaptured
argument. It determines whether captured pieces are returned in the ChessPieces
vector. Uses new pushPiece() private helper method.