Improve the ChessBoard class and related tests
Add a default ChessBoard constructor, initializes an empty board.

Add two setBoard() methods. One takes only a chessPieceSet argument and initializes
the board and the chess piece set to the normal initial positions. The other one takes
a chessPieceSet argument and two Positions vectors which are used to initialize the
board and chess piece set. These methods are called in the constructor.

Fix bug in isPositionValid() where it would only check the row number.

Add isPositionsValid() method which validates an array of positions. Used to validate
the Positions arguments to setBoard().

Add tests for the new constructors.